refactor(view): 상수 처리 진행
상수 처리 진행
