+++ src/main/java/racingcar/Application.java
@@ -1,12 +1,12 @@
 package racingcar;
 
-import racingcar.controller.RacingGameController;
+import racingcar.controller.RacingGameManager;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
 public class Application {
     public static void main(String[] args) {
-        RacingGameController racingGameController = new RacingGameController(
+        RacingGameManager racingGameController = new RacingGameManager(
                 new InputView(),
                 new OutputView()
         );
