+++ src/main/java/racingcar/view/InputView.java
@@ -4,13 +4,16 @@ import static camp.nextstep.edu.missionutils.Console.readLine;
 
 public class InputView {
 
+    private static final String INPUT_CAR_NAME_MESSAGE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
+    private static final String INPUT_TRY_COUNT_MESSAGE = "시도할 횟수는 몇 회인가요?";
+
     public String inputCarNames() {
-        System.out.println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)");
+        System.out.println(INPUT_CAR_NAME_MESSAGE);
         return readLine();
     }
 
     public String inputTryCount() {
-        System.out.println("시도할 횟수는 몇 회인가요?");
+        System.out.println(INPUT_TRY_COUNT_MESSAGE);
         return readLine();
     }
 }
