+++ src/main/java/racingcar/view/OutputView.java
@@ -5,8 +5,11 @@ import racingcar.domain.Car;
 
 public class OutputView {
 
+    private static final String ROUND_RESULT_HEADER = "실행 결과";
+    private static final String WINNER_HEADER = "최종 우승자 : ";
+
     public void printRoundHeader() {
-        System.out.println("실행 결과");
+        System.out.println(ROUND_RESULT_HEADER);
     }
 
     public void printRoundResult(List<Car> cars) {
@@ -15,9 +18,7 @@ public class OutputView {
         cars.forEach(car -> {
             stringBuilder.append(car.getName());
             stringBuilder.append(" : ");
-            for (int i = 0; i < car.getPosition(); i++) {
-                stringBuilder.append("-");
-            }
+            stringBuilder.append("-".repeat(Math.max(0, car.getPosition())));
             stringBuilder.append("\n");
         });
 
@@ -27,7 +28,7 @@ public class OutputView {
     public void printWinner(List<Car> winners) {
         StringBuilder stringBuilder = new StringBuilder();
 
-        stringBuilder.append("최종 우승자 : ");
+        stringBuilder.append(WINNER_HEADER);
 
         for (int i = 0; i < winners.size(); i++) {
             stringBuilder.append(winners.get(i).getName());
